Update eslint to v8.8.1 (#1561)

+++ src/app/portal/components/audios/audio-view/audio-view.component.ts
@@ -148,13 +148,14 @@ export class AudioViewComponent implements OnDestroy {
   }
 
   onDeleteConfirmed() {
-    this.transformedMedia.transformedMediaId &&
+    if (this.transformedMedia.transformedMediaId) {
       this.audioRequestService.deleteTransformedMedia(this.transformedMedia.transformedMediaId).subscribe({
         next: () => {
           this.router.navigate(['/audios']);
         },
         error: () => (this.isDeleting = false),
       });
+    }
   }
 
   onDownloadClicked() {
